Refactored random wheel level Removed Unused assets Started redesign of audio game
