+++ HamsterWheel/Assets.xcassets/Audio Game/RandomWheel/frameWButton.imageset/Contents.json
@@ -2,17 +2,15 @@
   "images" : [
     {
       "idiom" : "universal",
-      "filename" : "wheelFrame.png",
+      "filename" : "frameWButton.png",
       "scale" : "1x"
     },
     {
       "idiom" : "universal",
-      "filename" : "wheelFrame@2x.png",
       "scale" : "2x"
     },
     {
       "idiom" : "universal",
-      "filename" : "wheelFrame@3x.png",
       "scale" : "3x"
     }
   ],
