+++ HamsterWheel/AudioGame/RandomWheel/AGLevelSelector.swift
@@ -19,11 +19,6 @@ class AudioGameLevelSelector: SKScene {
             switch level {
             case 1: loadLevel1()
             case 2: loadLevel2()
-            case 3: loadLevel3()
-            case 4: loadLevel4()
-            case 5: loadLevel5()
-            case 6: loadLevel6()
-            case 7: loadLevel7()
                 
             default: returnToMainMenu()
             }
@@ -34,22 +29,10 @@ class AudioGameLevelSelector: SKScene {
     
     func loadLevel1() {
         if let view = view {
-            if let scene = SingleButtonAudioLevel(fileNamed: "AGButtonLevel") {
-                
+            if let scene = RandomWheel(fileNamed: "RandomWheel") {
                 scene.levelSelector = self
-                currentLevel = 1 // Just to be sure
-                // Level 1 variables
-                scene.animal = "Cow"
-                scene.defaultButtonImage = "cowButton"
-                scene.pressedButtonImage = "cowButtonPressed"
-                scene.audioFileName = "cowMoo"
-                scene.audioFileExtension = ".mp3"
-                scene.hasSpelling = false
-                
-                // Sets scale mode
+                currentLevel = 2 // Just to be sure
                 scene.scaleMode = .aspectFit
-                
-                // Present the scene
                 view.presentScene(scene)
             }
         }
@@ -58,131 +41,130 @@ class AudioGameLevelSelector: SKScene {
     func loadLevel2() {
         if let view = view {
             if let scene = SingleButtonAudioLevel(fileNamed: "AGButtonLevel") {
-                
+
                 scene.levelSelector = self
                 currentLevel = 2 // Just to be sure
-                
-                // Level 2 variables
-                scene.animal = "Dog"
-                scene.defaultButtonImage = "dogButton"
-                scene.pressedButtonImage = "dogButtonPressed"
-                scene.audioFileName = "dogBark"
-                scene.audioFileExtension = ".wav"
-                scene.hasSpelling = false
-                
-                // Sets scale mode
-                scene.scaleMode = .aspectFit
-                
-                // Present the scene
-                view.presentScene(scene)
-            }
-        }
-    }
-    
-    func loadLevel3() {
-        if let view = view {
-            if let scene = SingleButtonAudioLevel(fileNamed: "AGButtonLevel") {
-                
-                scene.levelSelector = self
-                currentLevel = 3 // Just to be sure
-                
-                // Level 3 variables
-                scene.animal = "Sheep"
-                scene.defaultButtonImage = "sheepButton"
-                scene.pressedButtonImage = "sheepButtonPressed"
-                scene.audioFileName = "sheepBaa"
-                scene.audioFileExtension = ".wav"
-                scene.hasSpelling = false
-                
-                // Sets scale mode
-                scene.scaleMode = .aspectFit
-                
-                // Present the scene
-                view.presentScene(scene)
-            }
-        }
-    }
-    
-    func loadLevel4() {
-        if let view = view {
-            if let scene = RandomWheel(fileNamed: "RandomWheel") {
-                scene.levelSelector = self
-                currentLevel = 4 // Just to be sure
-                scene.scaleMode = .aspectFit
-                view.presentScene(scene)
-            }
-        }
-    }
-    
-    func loadLevel5() {
-        if let view = view {
-            if let scene = SingleButtonAudioLevel(fileNamed: "AGButtonLevel") {
-                
-                scene.levelSelector = self
-                currentLevel = 5 // Just to be sure
-                
+
                 scene.animal = "Cow"
                 scene.defaultButtonImage = "cowButton"
                 scene.pressedButtonImage = "cowButtonPressed"
                 scene.audioFileName = "cowMoo"
                 scene.audioFileExtension = ".mp3"
                 scene.hasSpelling = true
-                
+
                 // Sets scale mode
                 scene.scaleMode = .aspectFit
-                
+
                 // Present the scene
                 view.presentScene(scene)
             }
         }
     }
     
-    func loadLevel6() {
-        if let view = view {
-            if let scene = SingleButtonAudioLevel(fileNamed: "AGButtonLevel") {
-                
-                scene.levelSelector = self
-                currentLevel = 6 // Just to be sure
-                
-                scene.animal = "Dog"
-                scene.defaultButtonImage = "dogButton"
-                scene.pressedButtonImage = "dogButtonPressed"
-                scene.audioFileName = "dogBark"
-                scene.audioFileExtension = ".wav"
-                scene.hasSpelling = true
-                
-                // Sets scale mode
-                scene.scaleMode = .aspectFit
-                
-                // Present the scene
-                view.presentScene(scene)
-            }
-        }
-    }
-    
-    func loadLevel7() {
-        if let view = view {
-            if let scene = SingleButtonAudioLevel(fileNamed: "AGButtonLevel") {
-                
-                scene.levelSelector = self
-                currentLevel = 7 // Just to be sure
-                
-                scene.animal = "Cat"
-                scene.defaultButtonImage = "catButton"
-                scene.pressedButtonImage = "catButtonPressed"
-                scene.audioFileName = "catMeow"
-                scene.audioFileExtension = ".wav"
-                scene.hasSpelling = true
-                
-                // Sets scale mode
-                scene.scaleMode = .aspectFit
-                
-                // Present the scene
-                view.presentScene(scene)
-            }
-        }
-    }
-    
+//    func loadLevel3() {
+//        if let view = view {
+//            if let scene = SingleButtonAudioLevel(fileNamed: "AGButtonLevel") {
+//
+//                scene.levelSelector = self
+//                currentLevel = 3 // Just to be sure
+//
+//                // Level 3 variables
+//                scene.animal = "Sheep"
+//                scene.defaultButtonImage = "sheepButton"
+//                scene.pressedButtonImage = "sheepButtonPressed"
+//                scene.audioFileName = "sheepBaa"
+//                scene.audioFileExtension = ".wav"
+//                scene.hasSpelling = false
+//
+//                // Sets scale mode
+//                scene.scaleMode = .aspectFit
+//
+//                // Present the scene
+//                view.presentScene(scene)
+//            }
+//        }
+//    }
+//
+//    func loadLevel4() {
+//        if let view = view {
+//            if let scene = RandomWheel(fileNamed: "RandomWheel") {
+//                scene.levelSelector = self
+//                currentLevel = 4 // Just to be sure
+//                scene.scaleMode = .aspectFit
+//                view.presentScene(scene)
+//            }
+//        }
+//    }
+//
+//    func loadLevel5() {
+//        if let view = view {
+//            if let scene = SingleButtonAudioLevel(fileNamed: "AGButtonLevel") {
+//
+//                scene.levelSelector = self
+//                currentLevel = 5 // Just to be sure
+//
+//                scene.animal = "Cow"
+//                scene.defaultButtonImage = "cowButton"
+//                scene.pressedButtonImage = "cowButtonPressed"
+//                scene.audioFileName = "cowMoo"
+//                scene.audioFileExtension = ".mp3"
+//                scene.hasSpelling = true
+//
+//                // Sets scale mode
+//                scene.scaleMode = .aspectFit
+//
+//                // Present the scene
+//                view.presentScene(scene)
+//            }
+//        }
+//    }
+//
+//    func loadLevel6() {
+//        if let view = view {
+//            if let scene = SingleButtonAudioLevel(fileNamed: "AGButtonLevel") {
+//
+//                scene.levelSelector = self
+//                currentLevel = 6 // Just to be sure
+//
+//                scene.animal = "Dog"
+//                scene.defaultButtonImage = "dogButton"
+//                scene.pressedButtonImage = "dogButtonPressed"
+//                scene.audioFileName = "dogBark"
+//                scene.audioFileExtension = ".wav"
+//                scene.hasSpelling = true
+//
+//                // Sets scale mode
+//                scene.scaleMode = .aspectFit
+//
+//                // Present the scene
+//                view.presentScene(scene)
+//            }
+//        }
+//    }
+//
+//    func loadLevel7() {
+//        if let view = view {
+//            if let scene = SingleButtonAudioLevel(fileNamed: "AGButtonLevel") {
+//
+//                scene.levelSelector = self
+//                currentLevel = 7 // Just to be sure
+//
+//                scene.animal = "Cat"
+//                scene.defaultButtonImage = "catButton"
+//                scene.pressedButtonImage = "catButtonPressed"
+//                scene.audioFileName = "catMeow"
+//                scene.audioFileExtension = ".wav"
+//                scene.hasSpelling = true
+//
+//                // Sets scale mode
+//                scene.scaleMode = .aspectFit
+//
+//                // Present the scene
+//                view.presentScene(scene)
+//            }
+//        }
+//    }
+
     func returnToMainMenu() {
         if let view = view {
             if let scene = SKScene(fileNamed: "MainMenuScene") {
