+++ HamsterWheel/AudioGame/RandomWheel/RandomWheel.swift
@@ -26,28 +26,18 @@ class RandomWheel: SKScene {
     var spinButton: UIButton?
     
     var audio: AVAudioPlayer?
-    var start: DispatchTime?
-    var end: DispatchTime?
-    var totalTime: Double?
     
     var menuButton: SKButton!
     var backButton: SKButton!
     var nextButton: SKButton!
     
     override func didMove(to view: SKView) {
-        // Adds the rotating wheel
-        addWheel(view: view, width: view.bounds.height * 0.8, height: view.bounds.height * 0.8)
-        
-        // Adds the spin button and the wheel frame
-        addWheelImages(view: view)
-        addSpinButton(view: view)
-        
-        // Start time in level
-        self.start = DispatchTime.now()
+        // Adds the TTWheel, the frame image and the spin button
+        addWheel(view: view)
         
         setupHomeButton()
         setupBackButton()
-        connectNextLevelButton()
+        setupNextLevelButton()
         
         // Fixes letter boxing on iPad
         sceneDidLayoutSubviews()
@@ -88,9 +78,6 @@ class RandomWheel: SKScene {
         /* Setup button selection handler for homescreen */
         backButton.selectedHandler = { [unowned self] in
             if let view = self.view {
-                // Calculates the time spend on the level
-                self.setEndTimeAndCalculateDifference()
-                
                 if let selector = self.levelSelector {
                     if selector.currentLevel != nil {
                         selector.currentLevel! -= 1
@@ -104,112 +91,20 @@ class RandomWheel: SKScene {
         }
     }
     
-    func connectNextLevelButton() {
+    func setupNextLevelButton() {
         nextButton = self.childNode(withName: "nextButton") as! SKButton
         nextButton.position = positionFromTop(CGPoint(x: 75.0, y: 275.0))
         nextButton.selectedHandler = transitionToNextScene
         nextButton.isHidden = true
     }
     
-    // MARK: Wheel
-    
-    // MARK: UI
-    
-    // Add both the wheel frame and spin button to the view
-    func addWheelImages(view: SKView) {
-        // Adds the wheel frame
-        frameImage = addImageToCenter(x: view.bounds.midX,
-                                      y: view.bounds.midY + 10,
-                                      height: view.bounds.height * 0.9 - 8,
-                                      width: view.bounds.height * 0.9,
-                                      image: #imageLiteral(resourceName: "wheelFrame"))
-        // adds the spin button image
-        spinButtonImage = addImageToCenter(x: view.bounds.midX,
-                                           y: view.bounds.midY + 5,
-                                           height: view.bounds.height * 0.3,
-                                           width: view.bounds.height * 0.3,
-                                           image: #imageLiteral(resourceName: "spin"))
-    }
-    
-    // Adds the rotating wheel to the view
-    func addWheel(view: SKView, width: CGFloat, height: CGFloat) {
-        // Define wheels frame
-        let wheelFrame = CGRect(x: view.bounds.midX, y: view.bounds.midY, width: width, height: height)
-        
-        // Initiate wheel
-        let fortuneWheel = TTFortuneWheel(frame: wheelFrame, slices: slices)
-        wheel = fortuneWheel
-        
-        // Set the center point of the wheel
-        fortuneWheel.center = CGPoint(x: view.bounds.midX, y: view.bounds.midY)
-        
-        // Sets all the slices to equal sizes
-        fortuneWheel.equalSlices = true
-        
-        // Has a black background by default so changing it to clear
-        fortuneWheel.backgroundColor = UIColor.clear
-        
-        // Setting individual styles on the slices
-        fortuneWheel.slices.enumerated().forEach { (pair) in
-            let slice = pair.element as! HWheelSlice
-            let offset = pair.offset
-            switch offset % 4 {
-            case 0: slice.style = .brickRed
-            case 1: slice.style = .sandYellow
-            case 2: slice.style = .babyBlue
-            case 3: slice.style = .deepBlue
-            default: slice.style = .brickRed
-            }
-        }
-        
-        // Add wheel to view
-        view.addSubview(fortuneWheel)
-    }
-    
-    func removeWheel() {
-        wheel?.removeFromSuperview()
-        frameImage?.removeFromSuperview()
-        spinButtonImage?.removeFromSuperview()
-        spinButton?.removeFromSuperview()
-    }
-    
-    // Adds the center Spin button to the view
-    func addSpinButton(view: SKView) {
-        let btn = UIButton()
-        btn.frame = CGRect(x: view.bounds.midX,
-                           y: view.bounds.midY + 10,
-                           width: 100,
-                           height: 100)
-        btn.center = CGPoint(x: view.bounds.midX,
-                             y: view.bounds.midY)
-        
-        
-        btn.setTitle("SPIN", for: .normal)
-        btn.setTitleColor(.yellow, for: .normal)
-        btn.titleLabel?.font = UIFont(name: "AmericanTypewriter-Bold", size: 20)
-        btn.addTarget(self, action: #selector(spinWheel), for: .touchUpInside)
-        spinButton = btn
-        view.addSubview(btn)
-    }
-    
-    // Adds an image to the view on the given location
-    func addImageToCenter(x: CGFloat, y: CGFloat, height: CGFloat, width: CGFloat, image: UIImage) -> UIImageView{
-        let imageView = UIImageView(frame: CGRect(x: x,
-                                                  y: y,
-                                                  width: width,
-                                                  height: height))
-        imageView.center = CGPoint(x: x, y: y)
-        imageView.image = image
-        imageView.contentMode = .scaleAspectFit
-        view?.addSubview(imageView)
-        return imageView
-    }
     // MARK: Functionality
     
     // Start the wheel spinning
     @objc func spinWheel() {
         if let wheel = wheel {
             wheel.startAnimating()
+            self.spinButton?.isEnabled = false
             DispatchQueue.main.asyncAfter(deadline: .now() + 1.5) {
                 let index = self.randomIndex()
                 self.wheel?.startAnimating(fininshIndex: index) { (finished) in
@@ -234,6 +129,7 @@ class RandomWheel: SKScene {
             case "Horse":   sound = "horseWhinnying";   extention = ".wav"
             default:        sound  = "cartoon_voice_says_yahoo"; extention = ".mp3"
             }
+            self.spinButton?.isEnabled = true
             self.playAudio(soundName: sound, soundExtention: extention)
         }
     }
@@ -255,9 +151,6 @@ class RandomWheel: SKScene {
     
     func transitionToNextScene() {
         if let view = view {
-            // Calculates the time spend on the level
-            setEndTimeAndCalculateDifference()
-            
             if let selector = levelSelector {
                 if selector.currentLevel != nil {
                     selector.currentLevel! += 1
@@ -269,29 +162,6 @@ class RandomWheel: SKScene {
                 view.presentScene(selector)
             }
         }
-        
-        //        // Calculates the time spend on the level
-        //        setEndTimeAndCalculateDifference()
-        //
-        //        // Creates and show next level
-        //        let level5 = AGlevel5(fileNamed: "AGlevel5")
-        //        level5?.scaleMode = .aspectFill
-        //
-        //        self.view?.presentScene(level5)
-        
-    }
-    
-    // FIXME: Add this function to end of level
-    // Sets the end time and calculates the time spent on the level using the start time
-    func setEndTimeAndCalculateDifference() {
-        // end time when level is complete
-        self.end = DispatchTime.now()
-        
-        // Difference in nano seconds (UInt64) converted to a Double
-        let nanoTime = Double((self.end?.uptimeNanoseconds)!) - Double((self.start?.uptimeNanoseconds)!)
-        let timeInterval = (nanoTime / 1000000000)
-        
-        self.totalTime = timeInterval
     }
     
     // Generates a random index based on the length of the slices array
